Adjust seed class names to required *TableSeeder format and reenable message module
