+++ modules/Message/module.json
@@ -3,7 +3,7 @@
     "alias": "message",
     "description": "",
     "keywords": [],
-    "active": 0,
+    "active": 1,
     "order": 2,
     "providers": [
         "Modules\\Message\\Providers\\MessageServiceProvider"
